Move more XContent parsers that are only used in tests to test codebase
Just like a couple times before, moving a couple more of the test only
parsers to the test codebase to save code-size etc.
