+++ server/src/main/java/org/elasticsearch/action/admin/cluster/settings/ClusterUpdateSettingsResponse.java
@@ -12,36 +12,19 @@ import org.elasticsearch.action.support.master.AcknowledgedResponse;
 import org.elasticsearch.common.io.stream.StreamInput;
 import org.elasticsearch.common.io.stream.StreamOutput;
 import org.elasticsearch.common.settings.Settings;
-import org.elasticsearch.xcontent.ConstructingObjectParser;
 import org.elasticsearch.xcontent.ParseField;
 import org.elasticsearch.xcontent.XContentBuilder;
-import org.elasticsearch.xcontent.XContentParser;
 
 import java.io.IOException;
 import java.util.Objects;
-
-import static org.elasticsearch.xcontent.ConstructingObjectParser.constructorArg;
 
 /**
  * A response for a cluster update settings action.
  */
 public class ClusterUpdateSettingsResponse extends AcknowledgedResponse {
 
-    private static final ParseField PERSISTENT = new ParseField("persistent");
-    private static final ParseField TRANSIENT = new ParseField("transient");
-
-    private static final ConstructingObjectParser<ClusterUpdateSettingsResponse, Void> PARSER = new ConstructingObjectParser<>(
-        "cluster_update_settings_response",
-        true,
-        args -> {
-            return new ClusterUpdateSettingsResponse((boolean) args[0], (Settings) args[1], (Settings) args[2]);
-        }
-    );
-    static {
-        declareAcknowledgedField(PARSER);
-        PARSER.declareObject(constructorArg(), (p, c) -> Settings.fromXContent(p), TRANSIENT);
-        PARSER.declareObject(constructorArg(), (p, c) -> Settings.fromXContent(p), PERSISTENT);
-    }
+    static final ParseField PERSISTENT = new ParseField("persistent");
+    static final ParseField TRANSIENT = new ParseField("transient");
 
     final Settings transientSettings;
     final Settings persistentSettings;
@@ -81,10 +64,6 @@ public class ClusterUpdateSettingsResponse extends AcknowledgedResponse {
         builder.startObject(TRANSIENT.getPreferredName());
         transientSettings.toXContent(builder, params);
         builder.endObject();
-    }
-
-    public static ClusterUpdateSettingsResponse fromXContent(XContentParser parser) {
-        return PARSER.apply(parser, null);
     }
 
     @Override
