+++ server/src/main/java/org/elasticsearch/action/admin/indices/open/OpenIndexResponse.java
@@ -11,8 +11,6 @@ package org.elasticsearch.action.admin.indices.open;
 import org.elasticsearch.action.support.master.ShardsAcknowledgedResponse;
 import org.elasticsearch.common.io.stream.StreamInput;
 import org.elasticsearch.common.io.stream.StreamOutput;
-import org.elasticsearch.xcontent.ConstructingObjectParser;
-import org.elasticsearch.xcontent.XContentParser;
 
 import java.io.IOException;
 
@@ -20,16 +18,6 @@ import java.io.IOException;
  * A response for a open index action.
  */
 public class OpenIndexResponse extends ShardsAcknowledgedResponse {
-
-    private static final ConstructingObjectParser<OpenIndexResponse, Void> PARSER = new ConstructingObjectParser<>(
-        "open_index",
-        true,
-        args -> new OpenIndexResponse((boolean) args[0], (boolean) args[1])
-    );
-
-    static {
-        declareAcknowledgedAndShardsAcknowledgedFields(PARSER);
-    }
 
     public OpenIndexResponse(StreamInput in) throws IOException {
         super(in, true);
@@ -43,9 +31,5 @@ public class OpenIndexResponse extends ShardsAcknowledgedResponse {
     public void writeTo(StreamOutput out) throws IOException {
         super.writeTo(out);
         writeShardsAcknowledged(out);
-    }
-
-    public static OpenIndexResponse fromXContent(XContentParser parser) {
-        return PARSER.apply(parser, null);
     }
 }
