+++ server/src/main/java/org/elasticsearch/action/admin/indices/validate/query/ValidateQueryResponse.java
@@ -9,22 +9,13 @@
 package org.elasticsearch.action.admin.indices.validate.query;
 
 import org.elasticsearch.action.support.DefaultShardOperationFailedException;
-import org.elasticsearch.action.support.broadcast.BaseBroadcastResponse;
 import org.elasticsearch.action.support.broadcast.BroadcastResponse;
-import org.elasticsearch.common.io.stream.StreamInput;
 import org.elasticsearch.common.io.stream.StreamOutput;
-import org.elasticsearch.xcontent.ConstructingObjectParser;
-import org.elasticsearch.xcontent.ParseField;
 import org.elasticsearch.xcontent.XContentBuilder;
-import org.elasticsearch.xcontent.XContentParser;
 
 import java.io.IOException;
-import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
-
-import static org.elasticsearch.xcontent.ConstructingObjectParser.constructorArg;
-import static org.elasticsearch.xcontent.ConstructingObjectParser.optionalConstructorArg;
 
 /**
  * The response of the validate action.
@@ -36,37 +27,9 @@ public class ValidateQueryResponse extends BroadcastResponse {
     public static final String VALID_FIELD = "valid";
     public static final String EXPLANATIONS_FIELD = "explanations";
 
-    @SuppressWarnings("unchecked")
-    static final ConstructingObjectParser<ValidateQueryResponse, Void> PARSER = new ConstructingObjectParser<>(
-        "validate_query",
-        true,
-        arg -> {
-            BaseBroadcastResponse response = (BaseBroadcastResponse) arg[0];
-            return new ValidateQueryResponse(
-                (boolean) arg[1],
-                (List<QueryExplanation>) arg[2],
-                response.getTotalShards(),
-                response.getSuccessfulShards(),
-                response.getFailedShards(),
-                Arrays.asList(response.getShardFailures())
-            );
-        }
-    );
-    static {
-        declareBroadcastFields(PARSER);
-        PARSER.declareBoolean(constructorArg(), new ParseField(VALID_FIELD));
-        PARSER.declareObjectArray(optionalConstructorArg(), QueryExplanation.PARSER, new ParseField(EXPLANATIONS_FIELD));
-    }
-
     private final boolean valid;
 
     private final List<QueryExplanation> queryExplanations;
-
-    ValidateQueryResponse(StreamInput in) throws IOException {
-        super(in);
-        valid = in.readBoolean();
-        queryExplanations = in.readCollectionAsList(QueryExplanation::new);
-    }
 
     ValidateQueryResponse(
         boolean valid,
@@ -114,9 +77,5 @@ public class ValidateQueryResponse extends BroadcastResponse {
             }
             builder.endArray();
         }
-    }
-
-    public static ValidateQueryResponse fromXContent(XContentParser parser) {
-        return PARSER.apply(parser, null);
     }
 }
