+++ server/src/main/java/org/elasticsearch/action/explain/ExplainResponse.java
@@ -17,14 +17,11 @@ import org.elasticsearch.core.RestApiVersion;
 import org.elasticsearch.index.get.GetResult;
 import org.elasticsearch.index.mapper.MapperService;
 import org.elasticsearch.rest.RestStatus;
-import org.elasticsearch.xcontent.ConstructingObjectParser;
 import org.elasticsearch.xcontent.ParseField;
 import org.elasticsearch.xcontent.ToXContentObject;
 import org.elasticsearch.xcontent.XContentBuilder;
-import org.elasticsearch.xcontent.XContentParser;
 
 import java.io.IOException;
-import java.util.Collection;
 import java.util.Objects;
 
 import static org.elasticsearch.common.lucene.Lucene.readExplanation;
@@ -35,14 +32,14 @@ import static org.elasticsearch.common.lucene.Lucene.writeExplanation;
  */
 public class ExplainResponse extends ActionResponse implements ToXContentObject {
 
-    private static final ParseField _INDEX = new ParseField("_index");
-    private static final ParseField _ID = new ParseField("_id");
+    static final ParseField _INDEX = new ParseField("_index");
+    static final ParseField _ID = new ParseField("_id");
     private static final ParseField MATCHED = new ParseField("matched");
-    private static final ParseField EXPLANATION = new ParseField("explanation");
-    private static final ParseField VALUE = new ParseField("value");
-    private static final ParseField DESCRIPTION = new ParseField("description");
-    private static final ParseField DETAILS = new ParseField("details");
-    private static final ParseField GET = new ParseField("get");
+    static final ParseField EXPLANATION = new ParseField("explanation");
+    static final ParseField VALUE = new ParseField("value");
+    static final ParseField DESCRIPTION = new ParseField("description");
+    static final ParseField DETAILS = new ParseField("details");
+    static final ParseField GET = new ParseField("get");
 
     private final String index;
     private final String id;
@@ -134,43 +131,6 @@ public class ExplainResponse extends ActionResponse implements ToXContentObject 
             out.writeBoolean(true);
             getResult.writeTo(out);
         }
-    }
-
-    private static final ConstructingObjectParser<ExplainResponse, Boolean> PARSER = new ConstructingObjectParser<>(
-        "explain",
-        true,
-        (arg, exists) -> new ExplainResponse((String) arg[0], (String) arg[1], exists, (Explanation) arg[2], (GetResult) arg[3])
-    );
-
-    static {
-        PARSER.declareString(ConstructingObjectParser.constructorArg(), _INDEX);
-        PARSER.declareString(ConstructingObjectParser.constructorArg(), _ID);
-        final ConstructingObjectParser<Explanation, Boolean> explanationParser = getExplanationsParser();
-        PARSER.declareObject(ConstructingObjectParser.optionalConstructorArg(), explanationParser, EXPLANATION);
-        PARSER.declareObject(ConstructingObjectParser.optionalConstructorArg(), (p, c) -> GetResult.fromXContentEmbedded(p), GET);
-    }
-
-    @SuppressWarnings("unchecked")
-    private static ConstructingObjectParser<Explanation, Boolean> getExplanationsParser() {
-        final ConstructingObjectParser<Explanation, Boolean> explanationParser = new ConstructingObjectParser<>(
-            "explanation",
-            true,
-            arg -> {
-                if ((float) arg[0] > 0) {
-                    return Explanation.match((float) arg[0], (String) arg[1], (Collection<Explanation>) arg[2]);
-                } else {
-                    return Explanation.noMatch((String) arg[1], (Collection<Explanation>) arg[2]);
-                }
-            }
-        );
-        explanationParser.declareFloat(ConstructingObjectParser.constructorArg(), VALUE);
-        explanationParser.declareString(ConstructingObjectParser.constructorArg(), DESCRIPTION);
-        explanationParser.declareObjectArray(ConstructingObjectParser.constructorArg(), explanationParser, DETAILS);
-        return explanationParser;
-    }
-
-    public static ExplainResponse fromXContent(XContentParser parser, boolean exists) {
-        return PARSER.apply(parser, exists);
     }
 
     @Override
