+++ server/src/main/java/org/elasticsearch/action/support/tasks/BaseTasksResponse.java
@@ -32,8 +32,8 @@ import static org.elasticsearch.ExceptionsHelper.rethrowAndSuppress;
  * Base class for responses of task-related operations
  */
 public class BaseTasksResponse extends ActionResponse {
-    protected static final String TASK_FAILURES = "task_failures";
-    protected static final String NODE_FAILURES = "node_failures";
+    public static final String TASK_FAILURES = "task_failures";
+    public static final String NODE_FAILURES = "node_failures";
 
     private List<TaskOperationFailure> taskFailures;
     private List<ElasticsearchException> nodeFailures;
