+++ server/src/main/java/org/elasticsearch/search/profile/SearchProfileResults.java
@@ -15,23 +15,15 @@ import org.elasticsearch.common.io.stream.StreamInput;
 import org.elasticsearch.common.io.stream.StreamOutput;
 import org.elasticsearch.common.io.stream.Writeable;
 import org.elasticsearch.core.Nullable;
-import org.elasticsearch.search.profile.aggregation.AggregationProfileShardResult;
-import org.elasticsearch.search.profile.query.QueryProfileShardResult;
 import org.elasticsearch.xcontent.ToXContentFragment;
 import org.elasticsearch.xcontent.XContentBuilder;
-import org.elasticsearch.xcontent.XContentParser;
 
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.Collections;
-import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 import java.util.TreeSet;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
-
-import static org.elasticsearch.common.xcontent.XContentParserUtils.ensureExpectedToken;
 
 /**
  * Profile results for all shards.
@@ -39,12 +31,12 @@ import static org.elasticsearch.common.xcontent.XContentParserUtils.ensureExpect
 public final class SearchProfileResults implements Writeable, ToXContentFragment {
 
     private static final Logger logger = LogManager.getLogger(SearchProfileResults.class);
-    private static final String ID_FIELD = "id";
+    public static final String ID_FIELD = "id";
     private static final String NODE_ID_FIELD = "node_id";
     private static final String CLUSTER_FIELD = "cluster";
     private static final String INDEX_NAME_FIELD = "index";
     private static final String SHARD_ID_FIELD = "shard_id";
-    private static final String SHARDS_FIELD = "shards";
+    public static final String SHARDS_FIELD = "shards";
     public static final String PROFILE_FIELD = "profile";
 
     // map key is the composite "id" of form [nodeId][(clusterName:)indexName][shardId] created from SearchShardTarget.toString
@@ -115,75 +107,6 @@ public final class SearchProfileResults implements Writeable, ToXContentFragment
     @Override
     public String toString() {
         return Strings.toString(this);
-    }
-
-    public static SearchProfileResults fromXContent(XContentParser parser) throws IOException {
-        XContentParser.Token token = parser.currentToken();
-        ensureExpectedToken(XContentParser.Token.START_OBJECT, token, parser);
-        Map<String, SearchProfileShardResult> profileResults = new HashMap<>();
-        while ((token = parser.nextToken()) != XContentParser.Token.END_OBJECT) {
-            if (token == XContentParser.Token.START_ARRAY) {
-                if (SHARDS_FIELD.equals(parser.currentName())) {
-                    while ((token = parser.nextToken()) != XContentParser.Token.END_ARRAY) {
-                        parseProfileResultsEntry(parser, profileResults);
-                    }
-                } else {
-                    parser.skipChildren();
-                }
-            } else if (token == XContentParser.Token.START_OBJECT) {
-                parser.skipChildren();
-            }
-        }
-        return new SearchProfileResults(profileResults);
-    }
-
-    private static void parseProfileResultsEntry(XContentParser parser, Map<String, SearchProfileShardResult> searchProfileResults)
-        throws IOException {
-        XContentParser.Token token = parser.currentToken();
-        ensureExpectedToken(XContentParser.Token.START_OBJECT, token, parser);
-        SearchProfileDfsPhaseResult searchProfileDfsPhaseResult = null;
-        List<QueryProfileShardResult> queryProfileResults = new ArrayList<>();
-        AggregationProfileShardResult aggProfileShardResult = null;
-        ProfileResult fetchResult = null;
-        String id = null;
-        String currentFieldName = null;
-        while ((token = parser.nextToken()) != XContentParser.Token.END_OBJECT) {
-            if (token == XContentParser.Token.FIELD_NAME) {
-                currentFieldName = parser.currentName();
-            } else if (token.isValue()) {
-                if (ID_FIELD.equals(currentFieldName)) {
-                    id = parser.text();
-                } else {
-                    parser.skipChildren();
-                }
-            } else if (token == XContentParser.Token.START_ARRAY) {
-                if ("searches".equals(currentFieldName)) {
-                    while ((parser.nextToken()) != XContentParser.Token.END_ARRAY) {
-                        queryProfileResults.add(QueryProfileShardResult.fromXContent(parser));
-                    }
-                } else if (AggregationProfileShardResult.AGGREGATIONS.equals(currentFieldName)) {
-                    aggProfileShardResult = AggregationProfileShardResult.fromXContent(parser);
-                } else {
-                    parser.skipChildren();
-                }
-            } else if (token == XContentParser.Token.START_OBJECT) {
-                if ("dfs".equals(currentFieldName)) {
-                    searchProfileDfsPhaseResult = SearchProfileDfsPhaseResult.fromXContent(parser);
-                } else if ("fetch".equals(currentFieldName)) {
-                    fetchResult = ProfileResult.fromXContent(parser);
-                } else {
-                    parser.skipChildren();
-                }
-            } else {
-                parser.skipChildren();
-            }
-        }
-        SearchProfileShardResult result = new SearchProfileShardResult(
-            new SearchProfileQueryPhaseResult(queryProfileResults, aggProfileShardResult),
-            fetchResult
-        );
-        result.getQueryPhase().setSearchProfileDfsPhaseResult(searchProfileDfsPhaseResult);
-        searchProfileResults.put(id, result);
     }
 
     /**
