+++ server/src/test/java/org/elasticsearch/action/admin/indices/validate/query/ValidateQueryResponseTests.java
@@ -10,18 +10,52 @@ package org.elasticsearch.action.admin.indices.validate.query;
 
 import org.elasticsearch.ElasticsearchException;
 import org.elasticsearch.action.support.DefaultShardOperationFailedException;
+import org.elasticsearch.action.support.broadcast.BaseBroadcastResponse;
 import org.elasticsearch.common.Strings;
 import org.elasticsearch.test.AbstractBroadcastResponseTestCase;
+import org.elasticsearch.xcontent.ConstructingObjectParser;
+import org.elasticsearch.xcontent.ParseField;
 import org.elasticsearch.xcontent.XContentParser;
 
 import java.io.IOException;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
+import static org.elasticsearch.action.support.broadcast.BaseBroadcastResponse.declareBroadcastFields;
+import static org.elasticsearch.xcontent.ConstructingObjectParser.constructorArg;
+import static org.elasticsearch.xcontent.ConstructingObjectParser.optionalConstructorArg;
+
 public class ValidateQueryResponseTests extends AbstractBroadcastResponseTestCase<ValidateQueryResponse> {
+
+    @SuppressWarnings("unchecked")
+    private static final ConstructingObjectParser<ValidateQueryResponse, Void> PARSER = new ConstructingObjectParser<>(
+        "validate_query",
+        true,
+        arg -> {
+            BaseBroadcastResponse response = (BaseBroadcastResponse) arg[0];
+            return new ValidateQueryResponse(
+                (boolean) arg[1],
+                (List<QueryExplanation>) arg[2],
+                response.getTotalShards(),
+                response.getSuccessfulShards(),
+                response.getFailedShards(),
+                Arrays.asList(response.getShardFailures())
+            );
+        }
+    );
+    static {
+        declareBroadcastFields(PARSER);
+        PARSER.declareBoolean(constructorArg(), new ParseField(ValidateQueryResponse.VALID_FIELD));
+        PARSER.declareObjectArray(
+            optionalConstructorArg(),
+            QueryExplanation.PARSER,
+            new ParseField(ValidateQueryResponse.EXPLANATIONS_FIELD)
+        );
+    }
 
     private static ValidateQueryResponse createRandomValidateQueryResponse(
         int totalShards,
@@ -60,7 +94,7 @@ public class ValidateQueryResponseTests extends AbstractBroadcastResponseTestCas
 
     @Override
     protected ValidateQueryResponse doParseInstance(XContentParser parser) throws IOException {
-        return ValidateQueryResponse.fromXContent(parser);
+        return PARSER.apply(parser, null);
     }
 
     @Override
