+++ server/src/test/java/org/elasticsearch/action/update/UpdateResponseTests.java
@@ -38,6 +38,7 @@ import static org.elasticsearch.action.DocWriteResponse.Result.NOT_FOUND;
 import static org.elasticsearch.action.DocWriteResponse.Result.UPDATED;
 import static org.elasticsearch.cluster.metadata.IndexMetadata.INDEX_UUID_NA_VALUE;
 import static org.elasticsearch.common.xcontent.XContentHelper.toXContent;
+import static org.elasticsearch.common.xcontent.XContentParserUtils.ensureExpectedToken;
 import static org.elasticsearch.test.XContentTestUtils.insertRandomFields;
 import static org.elasticsearch.test.hamcrest.ElasticsearchAssertions.assertToXContentEquivalent;
 
@@ -174,7 +175,7 @@ public class UpdateResponseTests extends ESTestCase {
         }
         UpdateResponse parsedUpdateResponse;
         try (XContentParser parser = createParser(xContentType.xContent(), mutated)) {
-            parsedUpdateResponse = UpdateResponse.fromXContent(parser);
+            parsedUpdateResponse = parseInstanceFromXContent(parser);
             assertNull(parser.nextToken());
         }
 
@@ -189,6 +190,32 @@ public class UpdateResponseTests extends ESTestCase {
         BytesReference parsedBytes = toXContent(parsedUpdateResponse, xContentType, humanReadable);
         BytesReference expectedBytes = toXContent(expectedUpdateResponse, xContentType, humanReadable);
         assertToXContentEquivalent(expectedBytes, parsedBytes, xContentType);
+    }
+
+    private static UpdateResponse parseInstanceFromXContent(XContentParser parser) throws IOException {
+        ensureExpectedToken(XContentParser.Token.START_OBJECT, parser.nextToken(), parser);
+
+        UpdateResponse.Builder context = new UpdateResponse.Builder();
+        while (parser.nextToken() != XContentParser.Token.END_OBJECT) {
+            parseXContentFields(parser, context);
+        }
+        return context.build();
+    }
+
+    /**
+     * Parse the current token and update the parsing context appropriately.
+     */
+    public static void parseXContentFields(XContentParser parser, UpdateResponse.Builder context) throws IOException {
+        XContentParser.Token token = parser.currentToken();
+        String currentFieldName = parser.currentName();
+
+        if (UpdateResponse.GET.equals(currentFieldName)) {
+            if (token == XContentParser.Token.START_OBJECT) {
+                context.setGetResult(GetResult.fromXContentEmbedded(parser));
+            }
+        } else {
+            DocWriteResponse.parseInnerToXContent(parser, context);
+        }
     }
 
     /**
