+++ server/src/test/java/org/elasticsearch/index/reindex/BulkByScrollTaskStatusTests.java
@@ -15,9 +15,14 @@ import org.elasticsearch.common.Randomness;
 import org.elasticsearch.common.Strings;
 import org.elasticsearch.common.io.stream.BytesStreamOutput;
 import org.elasticsearch.core.TimeValue;
+import org.elasticsearch.core.Tuple;
 import org.elasticsearch.index.reindex.BulkByScrollTask.Status;
 import org.elasticsearch.test.AbstractXContentTestCase;
+import org.elasticsearch.xcontent.ConstructingObjectParser;
+import org.elasticsearch.xcontent.ObjectParser;
+import org.elasticsearch.xcontent.ParseField;
 import org.elasticsearch.xcontent.ToXContent;
+import org.elasticsearch.xcontent.XContentParseException;
 import org.elasticsearch.xcontent.XContentParser;
 import org.hamcrest.Matchers;
 
@@ -32,6 +37,8 @@ import java.util.stream.IntStream;
 import static java.lang.Math.abs;
 import static java.util.stream.Collectors.toList;
 import static org.apache.lucene.tests.util.TestUtil.randomSimpleString;
+import static org.elasticsearch.common.xcontent.XContentParserUtils.ensureExpectedToken;
+import static org.elasticsearch.xcontent.ConstructingObjectParser.constructorArg;
 import static org.hamcrest.Matchers.equalTo;
 
 public class BulkByScrollTaskStatusTests extends AbstractXContentTestCase<BulkByScrollTask.Status> {
@@ -177,7 +184,142 @@ public class BulkByScrollTaskStatusTests extends AbstractXContentTestCase<BulkBy
 
     @Override
     protected BulkByScrollTask.Status doParseInstance(XContentParser parser) throws IOException {
-        return BulkByScrollTask.Status.fromXContent(parser);
+        XContentParser.Token token;
+        if (parser.currentToken() == XContentParser.Token.START_OBJECT) {
+            token = parser.nextToken();
+        } else {
+            token = parser.nextToken();
+        }
+        ensureExpectedToken(XContentParser.Token.START_OBJECT, token, parser);
+        token = parser.nextToken();
+        ensureExpectedToken(XContentParser.Token.FIELD_NAME, token, parser);
+        return innerParseStatus(parser);
+    }
+
+    private static final ConstructingObjectParser<Tuple<Long, Long>, Void> RETRIES_PARSER = new ConstructingObjectParser<>(
+        "bulk_by_scroll_task_status_retries",
+        true,
+        a -> new Tuple<>(((Long) a[0]), (Long) a[1])
+    );
+    static {
+        RETRIES_PARSER.declareLong(constructorArg(), new ParseField(BulkByScrollTask.Status.RETRIES_BULK_FIELD));
+        RETRIES_PARSER.declareLong(constructorArg(), new ParseField(BulkByScrollTask.Status.RETRIES_SEARCH_FIELD));
+    }
+
+    public static void declareFields(ObjectParser<? extends BulkByScrollTask.StatusBuilder, Void> parser) {
+        parser.declareInt(BulkByScrollTask.StatusBuilder::setSliceId, new ParseField(BulkByScrollTask.Status.SLICE_ID_FIELD));
+        parser.declareLong(BulkByScrollTask.StatusBuilder::setTotal, new ParseField(BulkByScrollTask.Status.TOTAL_FIELD));
+        parser.declareLong(BulkByScrollTask.StatusBuilder::setUpdated, new ParseField(BulkByScrollTask.Status.UPDATED_FIELD));
+        parser.declareLong(BulkByScrollTask.StatusBuilder::setCreated, new ParseField(BulkByScrollTask.Status.CREATED_FIELD));
+        parser.declareLong(BulkByScrollTask.StatusBuilder::setDeleted, new ParseField(BulkByScrollTask.Status.DELETED_FIELD));
+        parser.declareInt(BulkByScrollTask.StatusBuilder::setBatches, new ParseField(BulkByScrollTask.Status.BATCHES_FIELD));
+        parser.declareLong(
+            BulkByScrollTask.StatusBuilder::setVersionConflicts,
+            new ParseField(BulkByScrollTask.Status.VERSION_CONFLICTS_FIELD)
+        );
+        parser.declareLong(BulkByScrollTask.StatusBuilder::setNoops, new ParseField(BulkByScrollTask.Status.NOOPS_FIELD));
+        parser.declareObject(
+            BulkByScrollTask.StatusBuilder::setRetries,
+            RETRIES_PARSER,
+            new ParseField(BulkByScrollTask.Status.RETRIES_FIELD)
+        );
+        parser.declareLong(BulkByScrollTask.StatusBuilder::setThrottled, new ParseField(BulkByScrollTask.Status.THROTTLED_RAW_FIELD));
+        parser.declareFloat(
+            BulkByScrollTask.StatusBuilder::setRequestsPerSecond,
+            new ParseField(BulkByScrollTask.Status.REQUESTS_PER_SEC_FIELD)
+        );
+        parser.declareString(BulkByScrollTask.StatusBuilder::setReasonCancelled, new ParseField(BulkByScrollTask.Status.CANCELED_FIELD));
+        parser.declareLong(
+            BulkByScrollTask.StatusBuilder::setThrottledUntil,
+            new ParseField(BulkByScrollTask.Status.THROTTLED_UNTIL_RAW_FIELD)
+        );
+        parser.declareObjectArray(
+            BulkByScrollTask.StatusBuilder::setSliceStatuses,
+            (p, c) -> parseStatusOrException(p),
+            new ParseField(BulkByScrollTask.Status.SLICES_FIELD)
+        );
+    }
+
+    private static Status innerParseStatus(XContentParser parser) throws IOException {
+        XContentParser.Token token = parser.currentToken();
+        String fieldName = parser.currentName();
+        ensureExpectedToken(XContentParser.Token.FIELD_NAME, token, parser);
+        BulkByScrollTask.StatusBuilder builder = new BulkByScrollTask.StatusBuilder();
+        while ((token = parser.nextToken()) != XContentParser.Token.END_OBJECT) {
+            if (token == XContentParser.Token.FIELD_NAME) {
+                fieldName = parser.currentName();
+            } else if (token == XContentParser.Token.START_OBJECT) {
+                if (fieldName.equals(Status.RETRIES_FIELD)) {
+                    builder.setRetries(RETRIES_PARSER.parse(parser, null));
+                } else {
+                    parser.skipChildren();
+                }
+            } else if (token == XContentParser.Token.START_ARRAY) {
+                if (fieldName.equals(Status.SLICES_FIELD)) {
+                    while ((token = parser.nextToken()) != XContentParser.Token.END_ARRAY) {
+                        builder.addToSliceStatuses(parseStatusOrException(parser));
+                    }
+                } else {
+                    parser.skipChildren();
+                }
+            } else { // else if it is a value
+                switch (fieldName) {
+                    case Status.SLICE_ID_FIELD -> builder.setSliceId(parser.intValue());
+                    case Status.TOTAL_FIELD -> builder.setTotal(parser.longValue());
+                    case Status.UPDATED_FIELD -> builder.setUpdated(parser.longValue());
+                    case Status.CREATED_FIELD -> builder.setCreated(parser.longValue());
+                    case Status.DELETED_FIELD -> builder.setDeleted(parser.longValue());
+                    case Status.BATCHES_FIELD -> builder.setBatches(parser.intValue());
+                    case Status.VERSION_CONFLICTS_FIELD -> builder.setVersionConflicts(parser.longValue());
+                    case Status.NOOPS_FIELD -> builder.setNoops(parser.longValue());
+                    case Status.THROTTLED_RAW_FIELD -> builder.setThrottled(parser.longValue());
+                    case Status.REQUESTS_PER_SEC_FIELD -> builder.setRequestsPerSecond(parser.floatValue());
+                    case Status.CANCELED_FIELD -> builder.setReasonCancelled(parser.text());
+                    case Status.THROTTLED_UNTIL_RAW_FIELD -> builder.setThrottledUntil(parser.longValue());
+                }
+            }
+        }
+        return builder.buildStatus();
+    }
+
+    /**
+     * Since {@link BulkByScrollTask.StatusOrException} can contain either an {@link Exception} or a {@link Status} we need to peek
+     * at a field first before deciding what needs to be parsed since the same object could contains either.
+     * The {@link BulkByScrollTask.StatusOrException#EXPECTED_EXCEPTION_FIELDS} contains the fields that are expected when the serialised
+     * object was an instance of exception and the {@link Status#FIELDS_SET} is the set of fields expected when the
+     * serialized object was an instance of Status.
+     */
+    public static BulkByScrollTask.StatusOrException parseStatusOrException(XContentParser parser) throws IOException {
+        XContentParser.Token token = parser.currentToken();
+        if (token == null) {
+            token = parser.nextToken();
+        }
+        if (token == XContentParser.Token.VALUE_NULL) {
+            return null;
+        } else {
+            ensureExpectedToken(XContentParser.Token.START_OBJECT, token, parser);
+            token = parser.nextToken();
+            // This loop is present only to ignore unknown tokens. It breaks as soon as we find a field
+            // that is allowed.
+            while (token != XContentParser.Token.END_OBJECT) {
+                ensureExpectedToken(XContentParser.Token.FIELD_NAME, token, parser);
+                String fieldName = parser.currentName();
+                // weird way to ignore unknown tokens
+                if (Status.FIELDS_SET.contains(fieldName)) {
+                    return new BulkByScrollTask.StatusOrException(innerParseStatus(parser));
+                } else if (BulkByScrollTask.StatusOrException.EXPECTED_EXCEPTION_FIELDS.contains(fieldName)) {
+                    return new BulkByScrollTask.StatusOrException(ElasticsearchException.innerFromXContent(parser, false));
+                } else {
+                    // Ignore unknown tokens
+                    token = parser.nextToken();
+                    if (token == XContentParser.Token.START_OBJECT || token == XContentParser.Token.START_ARRAY) {
+                        parser.skipChildren();
+                    }
+                    token = parser.nextToken();
+                }
+            }
+            throw new XContentParseException("Unable to parse StatusFromException. Expected fields not found.");
+        }
     }
 
     @Override
