+++ x-pack/plugin/ent-search/src/main/java/org/elasticsearch/xpack/application/analytics/action/PutAnalyticsCollectionAction.java
@@ -146,20 +146,5 @@ public class PutAnalyticsCollectionAction {
             builder.field(COLLECTION_NAME_FIELD.getPreferredName(), name);
         }
 
-        private static final ConstructingObjectParser<Response, String> PARSER = new ConstructingObjectParser<>(
-            "put_analytics_collection_response",
-            false,
-            (p) -> {
-                return new Response((boolean) p[0], (String) p[1]);
-            }
-        );
-        static {
-            PARSER.declareString(constructorArg(), COLLECTION_NAME_FIELD);
-        }
-
-        public static Response fromXContent(String resourceName, XContentParser parser) throws IOException {
-            return new Response(AcknowledgedResponse.fromXContent(parser).isAcknowledged(), resourceName);
-        }
-
     }
 }
